Add CONTRIBUTORS
The CONTRIBUTORS file is an organize but tedious way to add the names of
the contributors of this project. If the .git is not maintained or is
removed upon making tarballs of this or for any other reasons, this file
will provide information about who contributed what into this project.
Plus, it adds a little fancyness >,<