Remove timestamp feature and improve D-Bus interfaces
For now, the timestamp feature seems a lot more harder to implement
especially with the datatypes being so different from D-Bus and Python.
This is why the timestamp feature will be inside a todo for now.

Wrapped functions that connect the program to D-Bus interfaces into a
function and that function is invoked every time there is a change or
error to the program, which may cause the D-Bus interfaces to
malfunction with outdated data.